+++ cypress/support/component-helper/constants.js
@@ -1,0 +1,21 @@
+export const SIZE = {
+  SMALL: "small",
+  MEDIUM: "medium",
+  LARGE: "large",
+};
+
+export const VALIDATION = {
+  ERROR: "rgb(203, 55, 74)",
+  WARNING: "rgb(239, 103, 0)",
+  INFO: "rgb(0, 96, 167)",
+};
+
+export const COLOR = {
+  BLACK: "rgb(0, 0, 0)",
+};
+
+export const CHARACTERS = {
+  DIACRITICS: "mp150ú¿¡üßä",
+  SPECIALCHARACTERS: "!@#$%^*()_+-=~[];:.,?{}&\"'<>",
+  STANDARD: "cypress_data",
+};