+++ src/components/dialog/dialog.style.ts
@@ -17,6 +17,7 @@ import {
   HORIZONTAL_PADDING,
   CONTENT_TOP_PADDING,
   CONTENT_BOTTOM_PADDING,
+  DialogSizes,
 } from "./dialog.config";
 import { ContentPaddingInterface } from "./dialog.component";
 
@@ -98,7 +99,7 @@ const calculatePaddingTopInnerContent = ({
 
 type StyledDialogProps = {
   topMargin: number;
-  size?: keyof typeof dialogSizes;
+  size?: DialogSizes;
   dialogHeight?: string;
 };
 
@@ -161,6 +162,7 @@ type StyledDialogTitleProps = {
 };
 
 const StyledDialogTitle = styled.div<StyledDialogTitleProps>`
+  background-color: var(--colorsUtilityYang100);
   padding: 23px ${HORIZONTAL_PADDING}px 0;
   border-bottom: 1px solid #ccd6db;
   ${({ showCloseIcon }) => showCloseIcon && "padding-right: 85px"};